Add Comment Function

+++ src/main/java/com/learning/springboot/controller/PublishController.java
@@ -39,7 +39,7 @@ public class PublishController {
      */
     @GetMapping("/publish/{id}")
     public String edit(
-            @PathVariable("id") Integer id,
+            @PathVariable("id") Long id,
             Model model){
         Question question = questionMapper.selectByPrimaryKey(id);
         model.addAttribute("id",id);
@@ -61,7 +61,7 @@ public class PublishController {
      */
     @PostMapping("/publish")
     public String doPublish(
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
